server/requirements.txt
Flask
Flask-Bcrypt
Flask-JWT-Extended
Flask-Cors
pymongo
python-dotenv